+++ app/scenes/Activity/index.js
@@ -31,7 +31,6 @@ import {
   setSummaryScreen,
   setSelected,
 } from "../../state/responses/responses.actions";
-import { setCumulativeActivities, setHiddenCumulativeActivities } from "../../state/activities/activities.actions";
 
 import { authTokenSelector } from "../../state/user/user.selectors";
 import ActivityScreens from "../../components/ActivityScreens";
@@ -50,7 +49,6 @@ import {
   isPrevEnabled,
 } from "../../services/activityNavigation";
 import Timer from "../../services/timer";
-import { evaluateCumulatives } from "../../services/scoring";
 
 const styles = StyleSheet.create({
   buttonArea: {
@@ -84,7 +82,7 @@ class Activity extends React.Component {
     this.setState({
       isSummaryScreen,
       idleTime,
-      isSplashScreen: activity.splash && activity.splash.en && currentScreen === 0
+      isSplashScreen: activity.splash && activity.splash.en && currentScreen === 0 && !isSummaryScreen
     }, () => {
       if (idleTime) {
         this.idleTimer.startCountdown(
@@ -130,36 +128,15 @@ class Activity extends React.Component {
       return;
     }
 
-    if (next === -1 && activity.compute && !isSummaryScreen) {
-      const { cumulativeActivities, hiddenCumulativeActivities, setCumulativeActivities, setHiddenCumulativeActivities } = this.props;
-      let { cumActivities } = evaluateCumulatives(responses, activity);
-      const cumulativeActivity = this.findActivity(cumActivities && cumActivities[0], currentApplet ?.activities) || {};
-
-      if (cumulativeActivities && cumulativeActivities[`${activity.id}/nextActivity`]) {
-        if (cumActivities.length > 0 && !hiddenCumulativeActivities ?.includes(activity.id)) setHiddenCumulativeActivities(activity.id);
-
-        cumActivities = _.difference(cumActivities, cumulativeActivities[`${activity.id}/nextActivity`]);
-        if (cumActivities.length > 0) {
-          cumActivities = [...cumulativeActivities[`${activity.id}/nextActivity`], ...cumActivities];
-          setCumulativeActivities({ [`${activity.id}/nextActivity`]: cumActivities });
-        }
-        if (hiddenCumulativeActivities ?.includes(cumulativeActivity ?.id)) setHiddenCumulativeActivities(cumulativeActivity ?.id, true);
-      } else {
-        setCumulativeActivities({ [`${activity.id}/nextActivity`]: cumActivities });
-        if (cumActivities.length > 0 && !hiddenCumulativeActivities ?.includes(activity.id))
-          setHiddenCumulativeActivities(activity.id);
-        if (hiddenCumulativeActivities ?.includes(cumulativeActivity ?.id)) setHiddenCumulativeActivities(cumulativeActivity ?.id, true);
-      }
-    }
-
     if ((autoAdvance || fullScreen) && !optionalText || goToNext) {
       if (next === -1 && activity.compute && !activity.summaryDisabled && !isSummaryScreen) {
         this.setState({ isSummaryScreen: true });
-        setSummaryScreen(true);
+        setSummaryScreen(activity, true);
       } else {
         if (isSummaryScreen) {
           this.setState({ isSummaryScreen: false });
         }
+
         nextScreen(timeElapsed);
         setSelected(false);
       }
@@ -208,11 +185,13 @@ class Activity extends React.Component {
       prevScreen,
       setSelected,
       isSelected,
+      currentResponse
     } = this.props;
+    const { activity, responses } = currentResponse;
 
     if (isSummaryScreen) {
       this.setState({ isSummaryScreen: false });
-      setSummaryScreen(false);
+      setSummaryScreen(activity, false);
       setSelected(false);
     } else {
       if (!currentScreen) {
@@ -270,11 +249,11 @@ class Activity extends React.Component {
       !isSummaryScreen
     ) {
       this.setState({ isSummaryScreen: true });
-      setSummaryScreen(true);
+      setSummaryScreen(activity, true);
     } else {
       if (isSummaryScreen) {
         this.setState({ isSummaryScreen: false });
-        setSummaryScreen(false);
+        setSummaryScreen(activity, false);
       }
       nextScreen();
       setSelected(false);
@@ -336,6 +315,7 @@ class Activity extends React.Component {
             title={activity.name.en}
             actionLabel={actionLabel}
             watermark={currentApplet.watermark}
+            isSummaryScreen={isSummaryScreen}
             prevLabel={prevLabel}
             topNavigation={topNavigation}
             prevEnabled={!isSummaryScreen && isPrevEnabled(currentScreen, activity)}
@@ -374,6 +354,7 @@ class Activity extends React.Component {
           <ActHeader
             title={activity.name.en}
             actionLabel={actionLabel}
+            isSummaryScreen={isSummaryScreen}
             watermark={currentApplet.watermark}
             topNavigation={topNavigation}
             prevEnabled={!isSummaryScreen && isPrevEnabled(currentScreen, activity)}
@@ -446,8 +427,6 @@ const mapStateToProps = (state) => ({
   itemVisibility: itemVisiblitySelector(state),
   isSummaryScreen: isSummaryScreenSelector(state),
   isSelected: state.responses.isSelected,
-  cumulativeActivities: state.activities.cumulativeActivities,
-  hiddenCumulativeActivities: state.activities.hiddenCumulativeActivities,
 });
 
 const mapDispatchToProps = {
@@ -459,8 +438,6 @@ const mapDispatchToProps = {
   completeResponse,
   setSummaryScreen,
   setActivitySelectionDisabled,
-  setCumulativeActivities,
-  setHiddenCumulativeActivities
 };
 
 export default connect(