+++ app/state/responses/responses.reducer.js
@@ -36,6 +36,8 @@ const replaceAppletResponses = (state, action) => {
 };
 
 export default (state = initialState, action = {}) => {
+  let activity, activityId;
+
   switch (action.type) {
     case RESPONSES_CONSTANTS.CLEAR:
       return initialState;
@@ -45,9 +47,18 @@ export default (state = initialState, action = {}) => {
         isSelected: action.payload,
       };
     case RESPONSES_CONSTANTS.SET_SUMMARYSCREEN:
+      activity = action.payload.activity;
+      activityId = activity.event ? activity.id + activity.event.id : activity.id;
+
       return {
         ...state,
-        isSummaryScreen: action.payload,
+        inProgress: {
+          ...state.inProgress,
+          [activityId]: {
+            ...state.inProgress[activityId],
+            isSummaryScreen: action.payload.isSummaryScreen
+          },
+        },
       };
     case RESPONSES_CONSTANTS.OPEN_ACTIVITY:
       return {
@@ -72,7 +83,7 @@ export default (state = initialState, action = {}) => {
         inProgress: R.omit([action.payload], state.inProgress),
       };
     case RESPONSES_CONSTANTS.CREATE_RESPONSE_IN_PROGRESS:
-      const { activity } = action.payload;
+      activity = action.payload.activity;
 
       return {
         ...state,
@@ -84,12 +95,15 @@ export default (state = initialState, action = {}) => {
             subjectId: action.payload.subjectId,
             timeStarted: action.payload.timeStarted,
             screenIndex: 0,
+            isSummaryScreen: false
           },
         },
       };
     case RESPONSES_CONSTANTS.SET_CURRENT_SCREEN:
-      const { activityId, screenIndex, startTime } = action.payload;
+      const { screenIndex, startTime } = action.payload;
       const { inProgress } = state;
+
+      activityId = action.payload.activityId;
 
       let time = {};
       if (activityId) {